+++ superset-frontend/src/katalon/DataTable/images/Table2.jpg
