+++ superset-frontend/src/katalon/DataTable/images/thumbnail.png
