+++ superset-frontend/src/katalon/DataTable/transformProps.ts
