+++ superset-frontend/src/katalon/DataTable/utils/formatValue.ts
