+++ superset-frontend/src/katalon/DataTable/utils/isEqualColumns.ts
