--- superset-frontend/src/katalon/chart-table/test/enzyme.tsx
@@ -1,67 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-import React, { ReactElement } from 'react';
-import { shallow as enzymeShallow, mount as enzymeMount } from 'enzyme';
-import {
-  EmotionCacheProvider,
-  createEmotionCache,
-  supersetTheme,
-  ThemeProvider,
-} from '@superset-ui/core';
-
-const emotionCache = createEmotionCache({
-  key: 'test',
-});
-
-type optionsType = {
-  wrappingComponentProps?: any;
-  wrappingComponent?: ReactElement;
-  context?: any;
-};
-
-export function ProviderWrapper(props: any) {
-  const { children, theme = supersetTheme } = props;
-  return (
-    <EmotionCacheProvider value={emotionCache}>
-      <ThemeProvider theme={theme}>{children}</ThemeProvider>
-    </EmotionCacheProvider>
-  );
-}
-
-export function mount(component: ReactElement, options: optionsType = {}) {
-  return enzymeMount(component, {
-    ...options,
-    wrappingComponent: ProviderWrapper,
-    wrappingComponentProps: {
-      theme: supersetTheme,
-      ...options?.wrappingComponentProps,
-    },
-  });
-}
-
-export function shallow(component: ReactElement, options: optionsType = {}) {
-  return enzymeShallow(component, {
-    ...options,
-    wrappingComponent: ProviderWrapper,
-    wrappingComponentProps: {
-      theme: supersetTheme,
-      ...options?.wrappingComponentProps,
-    },
-  }).dive();
-}
